+++ react-front-end/src/components/Sidebar.js
@@ -10,7 +10,7 @@ import Profile from "../components/Profile";
 import Footer from "../components/Footer";
 
 // images
-import ratingSecond from "../img/rating-second.png";
+import ratingSecond from "../img/level02.png";
 import vancSymbol from "../img/vancouver-symbol.jpeg";
 
 export default function Sidebar() {
